Fix of BigNumber conversions and Request updates
Added logic to convert all BigNumber references to Strings to avoid problems.
Added tweak that forced requests to be properly updated when approval count changed.

Still have issue with refreshing the page via the browser refresh. When this happens,
all the BigNumber references break.  Tried to console.log the values, but got a
Cross-Domain error issue.